demo2-python/inventory-service/requirements.txt
fastapi
uvicorn
motor
psycopg2-binary
sqlalchemy
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
python-dotenv
